ndg_commonmark::processor::extensions

Function process_manpage_references

pub fn process_manpage_references(
    html: &str,
    manpage_urls: Option<&HashMap<String, String>>,
) -> String

Process manpage references in HTML content.

This function processes manpage references by finding span elements with manpage-reference class and converting them to links when URLs are available.

Arguments

html - The HTML content to process

manpage_urls - Optional mapping of manpage names to URLs

Returns

The processed HTML with manpage references converted to links