ndg_commonmark::processor

Module core

Core implementation of the Markdown processor.

This module contains the main implementation of MarkdownProcessor and its methods, focused on the core rendering pipeline and configuration management.

Enums

ProcessorFeature

Features that can be queried on a processor instance.

Functions

collect_markdown_files

Collect all markdown files from the input directory

extract_inline_text

Extract all inline text from a heading node.