ndg_commonmark::syntax

Function create_default_manager

pub fn create_default_manager() -> SyntaxResult<SyntaxManager>

Create the default syntax manager based on available features.

This function will:

Use Syntastica if the ‘syntastica’ feature is enabled

Use Syntect if the ‘syntect’ feature is enabled

Return an error if both are enabled (mutual exclusivity check)

Return an error if neither is enabled

Note: While the Syntect feature is enabled, the two-face crate will also be pulled to provide additional Syntax highlighting.